refactored sortnet example
* split single main with many solver into Headers for the algorithm
  and 1 main per solver
* added benchmarks and cmake target benchmark_sortnets to run them
